Return primus instance from websocket service initialise method
Previously all uses of the websocket service used a shared primus instance which meant that the state of the websocket would be in an unpredictable state when running unit tests depending on the ordering of the tests.

Instead create a new primus instance per call to `initialise`, which means that the socket will always be in a fresh known state when running tests against it.