Refactor tag reset to use single action
Combining two actions to perform the reset was fragile since it was dependent on the combined invocation of the two functions to leave the state as desired. Combine the removal of existing tags and adding of the default tag into a single action to make it easier to unit test the state changes.