Fix formatting of default birthdays
The value returned by `Date#getMonth` is zero-indexed (see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/getMonth) so if we use the value to construct a YYYY-MM-DD date string then we need to add one to it to ensure a valid date is produced.

Currently date validation is failing because the dates returned are the 31st April.