Refactor websocket connection events to persist room id
When the websocket reconnects following a network loss it was reconnecting to a new room id, which prevents it from being able to receive search results from any in flight searches. Instead save the room id only once, and when the socket reconnects, re-subscribe to the same room id.

Additionally, ensure that the default tags are set only on the first connection and not on any subsequent reconnections.

Fixes #180