Remove tiles from `items` as well as `displayedItems`
The way our infinite scroll works we maintain two arrays of items, one of the items displayed (a.k.a `displayedItems`) and one of the complete item backlog (`items`).

When removing a tile by clicking the close button it was removing only from the `displayedItems` which meant that when triggering infinite scroll and fetching new items from the backlog then this still included the removed tile, causing it to be restored into the stack.

Fix this by removing the item from both arrays when triggering the `TILES_REMOVE_TILE` action.