Explicitly set `window.innerWidth` on test setup for isearch component
The rendering of the element is dependent on the window dimensions, so in order for the tests to be reliable we need to ensure that the dimensions are as expected before running the tests.

Tests are currently failing in phantomjs because it assumes a default window width of 400 pixels and the tests assume > 553.