Fix issues with tiles being lost
Because tiles is reset to be a filter of the mixture at https://github.com/numo-labs/isearch-ui/blob/master/src/actions/search-results.js#L187 any tiles which are not pushed onto the mixture are lost at the next iteration of the mixing function. This means that if there are more tiles than 1/6th of the number of packages then they are never rendered.

By ensuring that all tiles are always pushed onto the end of the mixture if they are not interspersed then we can prevent them from being lost.

Additionally remove the shuffle function, since we should now receive tiles in (roughly) ranking order.