Prevent isearch component triggering search on mount
Resetting the tags to default is already performed once the websocket connection is established and the ui is ready to receive results. Performing an additional reset here is redundant because the connection id is not yet present, and so cannot be used to pipe any results from the resultant search back to the user.