Tweak timing of loading search results to improve sorting
Show the first two results to be received immediately to give user feedback about the search success. Then either after 3 seconds, or a gap in result loading display the first 20.

This allows more of the early results to be buffered into the stream and therefore to be sorted using our ranking algorithm, while still giving the user fast feedback.