Set default departure date as part of the default state
Rather than defaulting it in the search details component, default it as part of the global default state.

This is needed because the departure date also needs to be available in order to render a weather tile. If it is only defaulted at a component level then it would need to be replicated twice. By setting it at a global level then it only needs to be set once.